* move modules to ntp-client/src/Network/NTP/Client * make demo application optional * add comment about ntpQuery
