+++ ntp-client/ntp-client.cabal
@@ -9,11 +9,16 @@ build-type:          Simple
 extra-source-files:  README.md
 cabal-version:       >=1.20
 
-Library
+flag demo
+  description: Build a demo application
+  default: True
+  manual: True
+
+library
   exposed-modules:      Network.NTP.Client
-                        Network.NTP.Packet
-                        Network.NTP.Query
-                        Network.NTP.Trace
+  other-modules:        Network.NTP.Client.Packet
+                        Network.NTP.Client.Query
+                        Network.NTP.Client.Trace
   build-depends:        async           >=2.2 && <2.3
                       , base            >=4.9 && <4.13
                       , binary          >=0.8 && <0.9
@@ -29,21 +34,26 @@ Library
   default-extensions:   GeneralizedNewtypeDeriving
                         
 test-suite test-ntp-client
-  hs-source-dirs:     test
+  hs-source-dirs:     test, src
   main-is:            Test.hs
+  other-modules:        Network.NTP.Client.Packet
   type:               exitcode-stdio-1.0
   build-depends:        base            >=4.9 && <4.13
                       , binary          >=0.8 && <0.9
+                      , time            >=1.6 && <1.10
                       , QuickCheck
                       , tasty
                       , tasty-quickcheck
-                      , ntp-client
   default-language:   Haskell2010
   ghc-options:        -Wall
 
 executable demo-ntp-client
   hs-source-dirs:     demo
   main-is:            Main.hs
+  if flag(demo)
+     buildable: True
+  else
+     buildable: False
   build-depends:        async           >=2.2 && <2.3
                       , base            >=4.9 && <4.13
                       , contra-tracer   >=0.1 && <0.2
